add posting board

+++ src/main/java/com/homepage/security/token/TokenRepository.java
@@ -14,7 +14,7 @@ public interface TokenRepository extends JpaRepository<Token, Integer>{
 			on t.user.id = u.id\s
 			where u.id = :id and (t.expired = false or t.revoked = false)\s
 			""")
-	List<Token> findAllValidTokenByUser(@Param("id") Integer id);
+	List<Token> findAllValidTokenByUser(@Param("id") Long id);
 	
 	Optional<Token> findByToken(String token);
 }
